# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    API for the WASAText application, which is the project for the WASA course @ Sapienza
  version: 0.0.1
...
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    message_id:
      name: message_id
      description: |
        ID of the message
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 4294967296
    comment_id:
      name: comment_id
      description: |
        ID of the comment
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 4294967296
  schemas:
    User:
      title: User
      description: |
        Schema defining a user, which can belong to multiple chats and send messages
      type: object
      properties:
        user_id:
          description: ID of the user
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: Maria
        display_name:
          description: Name of the user
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 32
          example: FairyMaria
        biography:
          description: |
            Little biography of the user
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 512
          example: When life gives you lemons...
        profile_pic:
          description: Profile picture
          type: string
          pattern: '^.*$'
          minLength: 0
          maxLength: 44294967296
          format: byte
      required:
        - user_id
        - display_name
        - biography
        - profile_pic
    Message:
      title: Message
      description: |
        Representation of a message
      type: object
      properties:
        message_id:
          description: ID of the message, always expressed in 16 digits
          type: integer
          example: 1
          minimum: 1
          maximum: 1048576 # Could be lower or higher, had to set a maximum
        sender:
          description: ID of the sender
          type: integer
          example: 2
          minimum: 1
          maximum: 2048
        chat:
          description: ID of the chat
          type: integer
          example: 1
          minimum: 1
          maximum: 65536
        content:
          description: Text of the message
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 65536
          example: Lorem ipsum dolor sit est cum magna gratia
        timestamp:
          description: Timestamp of when the message was sent
          type: string
          # format: date-time
          pattern: '^[0-9]*$'
          minLength: 0
          maxLength: 44294967296
          example: '1731321854'
        photo:
          description: Photo sent with the message, if any is sent. Can be null
          type: string
          format: byte
          pattern: '^.*$'
          minLength: 0
          maxLength: 44294967296
        received:
          description: List of users who received the message
          type: array
          minItems: 0
          maxItems: 44294967296
          items:
            $ref: "#/components/schemas/User"
        seen:
          description: List of users who read the message
          type: array
          minItems: 0
          maxItems: 44294967296
          items:
            $ref: "#/components/schemas/User"
        deleted:
          description: Whether the message has been deleted
          type: boolean
        forwarded:
          description: Specifies whether the message was forwarded
          type: boolean
          example: false
        reactions:
          description: Set of reactions applied to the message
          type: array
          minItems: 0
          maxItems: 44294967296
          items:
            $ref: "#/components/schemas/Reaction"
        reply:
          description: ID of the message to which the current replies. 0 means no reply
          type: integer
          example: 1
          minimum: 0
          maximum: 1048576
      required:
        - message_id
        - sender
        - content
        - chat
        - timestamp
        - photo
        - received
        - seen
        - deleted
        - forwarded
        - reactions
        - reply
    Reaction:
      title: Reaction
      description: |
        Defines a reaction, which can be made out of text or an emoji
      type: object
      properties:
        reaction_id:
          description: ID of the reaction
          type: integer
          example: 3
          minimum: 1
          maximum: 1048576
        message:
          description: Message ID to which the reaction corresponds
          type: integer
          example: 2
          minimum: 1
          maximum: 1048576
        sender:
          description: ID of the user
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: Maria
        content:
          description: Content of the reaction
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 16384
      required:
        - reaction_id
        - message
        - sender
        - content
...
security:
  - bearerAuth: []
paths:
...
  /message:
    # PUT = sendMessage

    put:
      tags: [ "Chats" ]
      summary: Send a message to the chat
      operationId: sendMessage
      description: Sends a message to the chat given the messageId
      requestBody:
        description: The message that must be sent
        content:
          application/json:
            schema:
              type: object
              description: A message to be sent. Is considered as primordial when sent through the API
              properties:
                sender:
                  description: ID of the sender
                  type: string
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  example: Maria
                chat:
                  description: ID of the chat where to send
                  type: integer
                  example: 4
                  minimum: 1
                  maximum: 65536
                content:
                  description: Text of the message
                  type: string
                  pattern: '^.*$'
                  minLength: 1
                  maxLength: 65536
                  example: Lorem ipsum dolor sit est cum magna gratia
                photo:
                  description: Photo sent with the message, if any is sent. Can be null
                  type: string
                  format: byte
                  pattern: '^.*$'
                  minLength: 0
                  maxLength: 44294967296
                replying:
                  description: ID of the message to which the user is replying
                  type: integer
                  example: 4
                  minimum: 0
                  maximum: 65536
        required: true
      responses:
        "200":
          description: |
            The message has been sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

        "418":
          description: |
            Couldn't send the message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /message/{message_id}:
    parameters:
      - $ref: "#/components/parameters/message_id"
    # DELETE = deleteMessage
    # POST = forwardMessage

    delete:
      tags: ["Chats"]
      summary: Delete a message from the chat
      operationId: deleteMessage
      description: Deletes a message from the chat, given the messageId
      responses:
        "200":
          description: |
            The message has been deleted successfully

        "400":
          description: |
            Couldn't delete the message

    post:
      tags: ["Chats"]
      summary: Forward a message to another chat
      description: Allows to forward a message to another chat
      operationId: forwardMessage
      requestBody:
        description: The chat to which the message should be forwarded
        content:
          application/json:
            schema:
              type: object
              properties:
                sender:
                  description: ID of the sender
                  type: string
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  example: Maria
                chat_id:
                  description: ID of the chat to which the message should be forwarded
                  type: integer
                  example: 4
                  minimum: 1
                  maximum: 65536
        required: true
      responses:
        "204":
          description: |
            The message has been forwarded to the chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

        "403":
          description: |
            Can't forward to another chat. Are you part of it?
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /comment:
    # PUT = commentMessage

    put:
      tags: ["Chats"]
      summary: Comment a message
      operationId: commentMessage
      description: Comments a message, given a messageId
      requestBody:
        description: The comment to add
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  description: Message ID to which the reaction corresponds
                  type: integer
                  example: 2
                  minimum: 1
                  maximum: 1048576
                sender:
                  description: ID of the user
                  type: string
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  example: Maria
                content:
                  description: Content of the reaction
                  type: string
                  pattern: '^.*$'
                  minLength: 1
                  maxLength: 16384
        required: true
      responses:
        "200":
          description: |
            The comment has been added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"

        "400":
          description: |
            There was an error while trying to add a reaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
  /comment/{comment_id}:
    parameters:
      - $ref: "#/components/parameters/comment_id"
    # DELETE = uncommentMessage

    delete:
      tags: ["Chats"]
      summary: Remove a comment from a message
      description: Removes a comment from a message, given the commentId
      operationId: uncommentMessage
      responses:
        "204":
          description: |
            The reaction has been removed successfully

        "400":
          description: |
            Couldn't delete the reaction. Are you part of the chat?